requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
requests==2.31.0
python-dotenv==1.0.0

# linting and typing
black==23.7.0
mypy==1.4.1
isort==5.12.0
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing_extensions==4.7.1
Flake8-pyproject==1.2.3
pylint==2.17.4